Add initial timer, it is currently not tied to the app-state
The timer currently does not do anything with the game state. It simply updates
every second and on a refresh/figwheel update will be reset back to zero. I saw
some weird issues trying to pull the timer state into a global atom (everytime
I saved the file and figwheel updated my browser, the timer updated faster) so
I am leaving it as is for now.

The timer was taken from the reagent examples at:
http://holmsand.github.io/reagent/