Enable touch, pull the :div.* tags out of our templates
Touch should now work. We have to grab all the elements with a 'grid-cell'
class and attach a 'touchstart' listener to them. This needs to happen
everytime the game grid is rendered. If we move to swapping cell locations
during a game where we attach the listeners (in init!) will need to be
rethought.

Pulled the :div.* keywords out so that it is easier to change them,
specifically the onces in our game grid rows.