Initial work at cleaning up the UI - new game panel at bottom
Eventually this will be where the target, score, timer and pause button live.
This change is part of the way there, although it is not clear what each of
them is doing.

The indiviual components also need to be cleaned up, the game panel is part of
the game component, and so the separate reagent components are no longer
necessary.